Abstract styling used by vitals table
