+++ packages/web/app/components/Table/index.js
@@ -9,3 +9,4 @@ export { TableFormFields } from './TableFormFields';
 export * from './TableButtonRow';
 export * from './useSelectableColumn';
 export * from './SearchTable';
+export { DynamicColumnTable } from './DynamicColumnTable';
